Refactor: move parse_reply to guard
