+++ tests/test_empty/components/test_parse_reply.py
@@ -1,6 +1,6 @@
 import unittest
 
-from trashcli.empty.parse_reply import parse_reply
+from trashcli.guard.parse_reply import parse_reply
 
 
 class TestParseReply(unittest.TestCase):
