+++ trashcli/guard/parse_reply.py
